requirements.txt
Django==4.2.3
djangorestframework==3.14.0
django-allauth==0.54.0
django-rest-auth==0.9.5
dj-rest-auth==4.0.1